Lesson4_Task1
round avatars with shadows
