+++ iOSLessonOne/VK_Lite_1/VK_Lite_1/Controllers/Cells/FriendCell.swift
@@ -15,11 +15,10 @@ class FriendCell: UITableViewCell {
     @IBOutlet weak var friendphotoImage: UIImageView!
     
     @IBOutlet weak var friendnameLabel: UILabel!
+    @IBOutlet weak var shadowView: ShadowViewController!
     
     override func awakeFromNib() {
         super.awakeFromNib()
-        
-        
     }
 }
