+++ iOSLessonOne/VK_Lite_1/VK_Lite_1/Controllers/FriendsTVController.swift
@@ -93,8 +93,6 @@ class FriendsTVController: UITableViewController {
         //let friendphotoImage =  UIImage(named:friends[indexPath.row].friendImageName)
         }
     }
-    
-
 }
