+++ iOSLessonOne/VK_Lite_1/VK_Lite_1/Controllers/Lecture4ViewController.swift
@@ -41,18 +41,18 @@ class Lecture4ViewController: UIViewController {
     
 }
 
-extension UIView {
-    
-    @IBInspectable var shadowColor: CGColor {
-        set { layer.shadowColor = newValue}
-        get { return layer.shadowColor ?? UIColor.black.cgColor}
-    }
-    @IBInspectable var shadowOpacity: Float {
-        set { layer.shadowOpacity = newValue}
-        get { return layer.shadowOpacity }
-    }
-    @IBInspectable var shadowRadius: CGFloat {
-        set { layer.shadowRadius = newValue}
-        get { return layer.shadowRadius}
-    }
-}
+//extension UIView {
+//    
+//    @IBInspectable var shadowColor: CGColor {
+//        set { layer.shadowColor = newValue}
+//        get { return layer.shadowColor ?? UIColor.black.cgColor}
+//    }
+//    @IBInspectable var shadowOpacity: Float {
+//        set { layer.shadowOpacity = newValue}
+//        get { return layer.shadowOpacity }
+//    }
+//    @IBInspectable var shadowRadius: CGFloat {
+//        set { layer.shadowRadius = newValue}
+//        get { return layer.shadowRadius}
+//    }
+//}
